Adding Asset identity (#53)
This commit adds a series of content identification methods to the images being loaded by the crawler. These include color, perceptual duplicates, url duplicates, OCR, page level RUM information, and more.

An identifier is a way to identify the content itself, or identify things about the content which can be surfaced in the modal and/or the reports.

These identification methods are pluggable so that they can be added and removed easily as we explore options with our customers.

This commit makes the report download pluggable for various types of reports, and adds an experimental assets success score to the report.

It brings in several dependencies to perform these actions, which you can see in the index.html.